update struct, adding mock struct
