+++ registryadaptors/armosec/v1/datastructures.go
@@ -1,4 +1,4 @@
-package containerimages
+package v1
 
 import "time"
 
@@ -25,4 +25,34 @@ type V2ListRequest struct {
 	// Don't expose FieldsList outside without well designed decision
 	FieldsList              []string          `json:"includeFields,omitempty"`
 	FieldsReverseKeywordMap map[string]string `json:"-,omitempty"`
+}
+
+type FeLoginData struct {
+	Secret   string `json:"secret"`
+	ClientId string `json:"clientId"`
+}
+
+type FeLoginResponse struct {
+	Token        string `json:"accessToken"`
+	RefreshToken string `json:"refreshToken"`
+	ExpiresIn    int32  `json:"expiresIn"`
+	Expires      string `json:"expires"`
+}
+
+type ArmoBeConfiguration struct {
+	BackendUrl string `json:"backend"`
+	AuthUrl    string `json:"authUrl"`
+}
+type ArmoSelectCustomer struct {
+	SelectedCustomerGuid string `json:"selectedCustomer"`
+}
+
+type ArmoCivAdaptor struct {
+	registry   string
+	accountId  string
+	clientId   string
+	accessKey  string
+	feToken    FeLoginResponse
+	armoUrls   ArmoBeConfiguration
+	authCookie string
 }
